backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.96.0"
uvicorn = "^0.22.0"
motor = "^3.1.2"
python-dotenv = "^1.0.0"
email-validator = "^2.0.0.post2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
httpx = {extras = ["cli"], version = "^0.24.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
